refactor: rid of extra project-permisions.model file gf-557
