+++ apps/backend/src/modules/project-permissions/project-permissions.ts
@@ -17,3 +17,4 @@ const projectPermissionsController = new ProjectPermissionsController(
 );
 
 export { projectPermissionsController };
+export { ProjectPermissionModel } from "./project-permissions.model.js";
